LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/top-layer-nested-dialog-expected.txt


FAIL Dialogs initially not matching for container queries assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 255, 0)"
FAIL Dialogs still not matching after showModal assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 255, 0)"
PASS @container queries start matching
